Use the right type for attribute binding location.

+++ src/AbstractShaderProgram.h
@@ -203,13 +203,13 @@ class MAGNUM_EXPORT AbstractShaderProgram {
         /**
          * @brief Base struct for attribute location and type
          *
-         * See AbstractShaderProgram documentation or Mesh::bindAttribute()
+         * See @ref AbstractShaderProgram-subclassing or Mesh::bindAttribute()
          * for an example.
          *
          * @todo Support for BGRA attribute type (OpenGL 3.2, @extension{ARB,vertex_array_bgra})
          */
-        template<size_t i, class T> struct Attribute {
-            static const size_t Location = i;   /**< Location to which the attribute is bound */
+        template<GLuint i, class T> struct Attribute {
+            static const GLuint Location = i;   /**< Location to which the attribute is bound */
             typedef T Type;                     /**< %Attribute type */
         };
